refactor: consolidate project switcher logic into useProjectSwitcher hook
- Move task gathering, grouping, and navigation into useProjectSwitcher
- Simplify modal to consume the hook instead of inline logic
- Remove get-switcher-tasks.ts (absorbed into the hook)
- Ctrl+Tab directly navigates to next/prev task without opening modal
